refactor: Organizados namespaces, repassando DTOs de requisição e resposta do UseCase para camada de aplicação
