+++ CleanArchitecture.ISP/Application/OperationResult/Result.cs
@@ -1,8 +1,7 @@
-﻿using CleanArchitecture.ISP.Domain.PointClocks.Actions.Biometrics.OperationResult;
-using CleanArchitecture.ISP.Domain.PointClocks.Actions.PointClockDateTime.OperationResult;
+﻿using CleanArchitecture.ISP.Application.UseCases;
 using System.Text.Json.Serialization;
 
-namespace CleanArchitecture.ISP.Domain.OperationResult;
+namespace CleanArchitecture.ISP.Application.OperationResult;
 
 [JsonDerivedType(typeof(SetDateTimeSuccess), nameof(SetDateTimeSuccess))]
 [JsonDerivedType(typeof(ShareBiometricsSuccess), nameof(ShareBiometricsSuccess))]
