--- CleanArchitecture.ISP/Domain/PointClocks/Actions/Biometrics/DTOs/ShareBiometricsRequest.cs
@@ -1,6 +1,0 @@
-﻿using CleanArchitecture.ISP.Domain.PointClocks.Entities;
-
-namespace CleanArchitecture.ISP.Domain.PointClocks.Actions.Biometrics.DTOs;
-public record ShareBiometricsRequest(PointClock[] PointClocks)
-{
-}
