--- CleanArchitecture.ISP/Domain/PointClocks/Actions/PointClockDateTime/DTOs/SetDateTimeRequest.cs
@@ -1,6 +1,0 @@
-﻿using CleanArchitecture.ISP.Domain.PointClocks.Entities;
-
-namespace CleanArchitecture.ISP.Domain.PointClocks.Actions.PointClockDateTime.DTOs;
-public record SetDateTimeRequest(PointClock[] PointClocks)
-{
-}
