+++ CleanArchitecture.ISP/Domain/PointClocks/Gateway/IControlIDGateway.cs
@@ -1,4 +1,4 @@
-﻿using CleanArchitecture.ISP.Domain.OperationResult;
+﻿using CleanArchitecture.ISP.Application.OperationResult;
 
 namespace CleanArchitecture.ISP.Domain.PointClocks.Gateway;
 public interface IControlIDGateway
