+++ CleanArchitecture.ISP/Infrastructure/PointClockGateways/ControlIDGateway.cs
@@ -1,7 +1,5 @@
-﻿using CleanArchitecture.ISP.Application.UseCases;
-using CleanArchitecture.ISP.Domain.OperationResult;
-using CleanArchitecture.ISP.Domain.PointClocks.Actions.Biometrics.OperationResult;
-using CleanArchitecture.ISP.Domain.PointClocks.Actions.PointClockDateTime.OperationResult;
+﻿using CleanArchitecture.ISP.Application.OperationResult;
+using CleanArchitecture.ISP.Application.UseCases;
 using CleanArchitecture.ISP.Domain.PointClocks.Entities;
 using CleanArchitecture.ISP.Domain.PointClocks.Gateway;
